ファイル名、ID、クラス名の変更とコメントアウトの削除
