--- iOS-study/NewViewController.swift
@@ -1,35 +1,0 @@
-//
-//  NewViewController.swift
-//  iOS-study
-//
-//  Created by S892 on 2023/06/23.
-//
-
-import UIKit
-
-class NewViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
-    }
-    
-    override func viewDidAppear(_ animated: Bool){
-        super.viewDidAppear(animated)
-
-        if let oldViewController = self.storyboard?.instantiateViewController(withIdentifier: "ViewController") {
-            self.present(oldViewController, animated: true, completion: nil)
-        }
-    }
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destination.
-        // Pass the selected object to the new view controller.
-    }
-    */
-
-}
