+++ iOS-study/WeatherViewController.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 
-class ViewController: UIViewController {
+class WeatherViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
